+++ src/Hooks.php
@@ -16,8 +16,11 @@ use Kebabble\Config\Registration;
 use Kebabble\Config\Settings;
 use Kebabble\Config\OrderFields;
 use Kebabble\Config\CompanyFields;
+use Kebabble\API\Mention;
 
 use WP_Term;
+use WP_Error;
+use WP_REST_Request;
 
 /**
  * Handles and translates the hooks into kebabble class functions.
@@ -73,6 +76,13 @@ class Hooks {
 	protected $save;
 
 	/**
+	 * Handles mentions from and to the Slack API.
+	 *
+	 * @var Mention
+	 */
+	protected $api_mention;
+
+	/**
 	 * Constructor.
 	 *
 	 * @param Registration  $registration   Connects with the system required objects.
@@ -82,8 +92,9 @@ class Hooks {
 	 * @param OrderFields   $order_fields   Fields for the order form.
 	 * @param CompanyFields $company_fields Fields for the company taxonomy.
 	 * @param Save          $save           Storage processing for saved term items.
+	 * @param Mention       $api_mention    Handles mentions from and to the Slack API.
 	 */
-	public function __construct( Registration $registration, Publish $publish, Delete $delete, Settings $settings, OrderFields $order_fields, CompanyFields $company_fields, Save $save ) {
+	public function __construct( Registration $registration, Publish $publish, Delete $delete, Settings $settings, OrderFields $order_fields, CompanyFields $company_fields, Save $save, Mention $api_mention ) {
 		$this->registration   = $registration;
 		$this->publish        = $publish;
 		$this->delete         = $delete;
@@ -91,6 +102,7 @@ class Hooks {
 		$this->order_fields   = $order_fields;
 		$this->company_fields = $company_fields;
 		$this->save           = $save;
+		$this->api_mention    = $api_mention;
 	}
 
 	/**
@@ -113,7 +125,7 @@ class Hooks {
 		add_action( 'admin_enqueue_scripts', [ &$this, 'enqueued_scripts' ] );
 
 		// Order functionality.
-		add_action( 'publish_kebabble_orders', [ &$this->publish, 'handle_publish' ], 10, 2 );
+		add_action( 'publish_kebabble_orders', [ &$this->publish, 'hook_handle_publish' ], 10, 2 );
 		add_filter( 'wp_insert_post_data', [ &$this->publish, 'change_title' ], 99, 2 );
 		add_action( 'trash_kebabble_orders', [ &$this->delete, 'handle_deletion' ], 10, 2 );
 		add_action( 'untrash_post', [ &$this->delete, 'handle_undeletion' ], 10, 2 );
@@ -121,6 +133,9 @@ class Hooks {
 		// Company functionality.
 		add_action( 'created_kebabble_company', [ &$this->save, 'save_custom_company_details' ] );
 		add_action( 'edited_kebabble_company', [ &$this->save, 'save_custom_company_details' ] );
+
+		// API Hooks.
+		add_action( 'rest_api_init', [ &$this, 'api_endpoints' ] );
 	}
 
 	/**
@@ -135,4 +150,27 @@ class Hooks {
 		}
 	}
 
+	/**
+	 * Registers the Kebabble API endpoints.
+	 *
+	 * @return void
+	 */
+	public function api_endpoints():void {
+		// ?rest_route=/kebabble/v1/test
+		// Just a comical test message, as I've had problems with the API before (permalinks, normally).
+		register_rest_route( 'kebabble/v1', '/test', [
+			'methods'  => 'GET',
+			'callback' => function ( WP_REST_Request $r ) {
+				error_log( wp_json_encode( $r ) );
+				return new WP_Error( 'kebabble_called_test', "I'm a little teapot, short and stout. Here's my handle, here's my spout. When I see the tea cups hear me shout. Tip me up and pour me out.", [ 'status' => 418 ] );
+			},
+		]);
+
+		// ?rest_route=/kebabble/v1/slack
+		// All Slack-based comms via the Events API are sent to only one endpoint.
+		register_rest_route( 'kebabble/v1', 'slack', [
+			'methods'  => 'POST',
+			'callback' => [ &$this->api_mention, 'main' ],
+		]);
+	}
 }