+++ src/Library/Slack.php
@@ -16,81 +16,36 @@ use Carbon\Carbon;
 
 /**
  * Intermediary between Kebabble WP and Slack.
+ *
+ * @todo Become parent of botclient. Remove formatting dependency and declare in use.
  */
-class Slack {
-	/**
-	 * Pre-processing before publishing.
-	 *
-	 * @var Formatting
-	 */
-	public $formatting;
-
-	/**
-	 * Slack API communication handler.
-	 *
-	 * @var Slack
-	 */
-	public $slack;
-
+class Slack extends botclient {
 	/**
 	 * Constructor.
 	 *
-	 * @param Formatting $formatting Pre-processing before publishing.
+	 * @param string $channel Optional channel override.
 	 */
-	public function __construct( Formatting $formatting ) {
-		$this->formatting = $formatting;
-		$this->slack      = $this->generate_slackbot();
+	public function __construct( ?string $channel = null ) {
+		parent::__construct(
+			get_option( 'kbfos_settings' )['kbfos_botkey'],
+			( empty( $channel ) ) ? get_option( 'kbfos_settings' )['kbfos_botchannel'] : $channel
+		);
 	}
 
 	/**
-	 * Sends contents over to the Slack message API.
+	 * Sends a message to the Slack channel.
 	 *
-	 * @param integer     $id                 Post ID.
-	 * @param array       $order              Order array to be displayed on Slack.
+	 * @param string      $message            Desired message to be displayed.
 	 * @param string|null $existing_timestamp If an existing timestamp is passed, that message is modified.
 	 * @param string|null $override_channel   Change the Slack channel, if desired.
+	 * @param string|null $thread_timestamp   Timestamp of a thread, if desired.
 	 * @return string Unique timestamp of the message, used for editing.
 	 */
-	public function send_to_slack( int $id, array $order, ?string $existing_timestamp = null, ?string $override_channel = null ):string {
-		$so = $this->slack;
-		if ( ! empty( $override_channel ) ) {
-			$so = $this->generate_slackbot( $override_channel );
-		}
-
-		$timestamp = null;
-		if ( $order['override']['enabled'] ) {
-			// Custom Message.
-			$timestamp = $so->message( $order['override']['message'], ( ! empty( $existing_timestamp ) ) ? $existing_timestamp : false );
-		} else {
-			// Generated message.
-			$timestamp = $so->message(
-				$this->formatting->status(
-					$id,
-					$order['food'],
-					$order['order'],
-					$order['driver'],
-					(int) $order['tax'],
-					Carbon::parse( get_the_date( 'Y-m-d H:i:s', $id ) ),
-					( is_array( $order['payment'] ) ) ? $order['payment'] : [ $order['payment'] ],
-					$order['paymentLink']
-				),
-				( ! empty( $existing_timestamp ) ) ? $existing_timestamp : false
-			);
-		}
-
-		return $timestamp;
-	}
-
-	/**
-	 * Sets up the Slack bot client object.
-	 *
-	 * @param string $custom_channel Override the default option channel.
-	 * @return botclient
-	 */
-	private function generate_slackbot( string $custom_channel = '' ):botclient {
-		return new botclient(
-			get_option( 'kbfos_settings' )['kbfos_botkey'],
-			( empty( $custom_channel ) ) ? get_option( 'kbfos_settings' )['kbfos_botchannel'] : $custom_channel
+	public function send_message( string $message, ?string $existing_timestamp = null, ?string $override_channel = null, ?string $thread_timestamp = null ):string {
+		return $this->message(
+			$message,
+			( ! empty( $existing_timestamp ) ) ? $existing_timestamp : false,
+			( ! empty( $thread_timestamp ) ) ? $thread_timestamp : false
 		);
 	}
 }